requirements.txt
anyio==3.6.1
asgiref==3.5.2
attrs==21.4.0
bcrypt==3.2.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
distro==1.7.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
fastapi==0.70.0
h11==0.13.0
idna==3.3
iniconfig==1.1.1
jsonschema==3.2.0
packaging==21.3
paramiko==2.11.0
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
pydantic==1.8.2
PyNaCl==1.5.0
pyparsing==3.0.8
pyrsistent==0.18.1
pytest==7.1.2
pytest-docker==0.12.0
python-dotenv==0.20.0
PyYAML==5.4.1
requests==2.27.1
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
texttable==1.6.4
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
websocket-client==0.59.0
